Add explicit unsigned short casting where necessary

+++ src/models/Constants.h
@@ -18,11 +18,11 @@ namespace models {
         static const unsigned short width = 10;
         static const unsigned short height = 10;
         static const ShipTypes shipTypes = {
-                {"Speed boat", 2},
-                {"Destroyer", 3},
-                {"Submarine", 3},
-                {"Battleship", 4},
-                {"Aircraft Carrier", 5},
+                {"Speed boat", (unsigned short) 2},
+                {"Destroyer", (unsigned short) 3},
+                {"Submarine", (unsigned short) 3},
+                {"Battleship", (unsigned short) 4},
+                {"Aircraft Carrier", (unsigned short) 5},
         };
         static const std::vector<std::string> VALID_PLAYER_NAMES = {"Richard", "Steve"};
     }
